refactor : 체스 게임 룰에 따라 체스 장기말을 세팅하는 작업을 별도의 클래스에 위임
