+++ src/test/java/domain/BoardTest.java
@@ -13,7 +13,7 @@ class BoardTest {
     @Test
     @DisplayName("체스판의 현재 상태를 조회할 수 있다.")
     void initialize() {
-        Board board = new Board();
+        Board board = Board.initialize();
         List<List<Piece>> status = board.findCurrentStatus();
 
         assertThat(status).hasSize(8);
